bspwm: Add basic restarting functionality
Add very basic restarting functionality that will reload all of the
configurations. There are a lot of problems still, namely that windows
that were already open are left in a weird unmanaged state and certain
applications could be started again without being stopped from before.

A mechanism to avoid this has been added to the wrapper scrript, but
the bspwmr doesn't use it yet.